Refactor member service by taking extracting some methods
